Switch to YAML-tmLanguage files
These are easier to write, and there is no difference in the end because
they are converted to PList format, same as the JSON-tmLanguage files.

We clean up the file a bit since YAML is more expressive and allows
comments in different places. (This has the side effect of removing the
comments from the PList file, since the YAML comments are not converted.
This is OK, because we treat the PList file as machine-readable only.)

We also add PackageDev markers to the files, in order to make it easier
to use PackageDev's "Convert" command to regenerate the PList files
without having to select the target format.